getAllUser done

+++ src/main/java/com/example/banknator/users/dto/UpdateUserInformation.java
@@ -3,13 +3,13 @@ package com.example.banknator.users.dto;
 import java.time.LocalDate;
 
 public record UpdateUserInformation(
-    Integer id,
+    long id,
     String firstName,
     String lastName,
     String email,
     String password,
     String address,
-    Long phone,
+    long phone,
     Integer creditScore,
     LocalDate dateOfBirth
 ) {
